fix: message board complete fix
The updating of the message board with log messages has been fixed by using the data tracker to denote the button that has been clicked. Still has some bugs that arise from the not perfect location of the buttons. Also, it should be checked if this method can be modified as the message board is now a child of the GearVRRig and that is not desirable.